tests/wpt/web-platform-tests/workers/interfaces/WorkerUtils/importScripts/007.html

<!--
importScripts('data:text/javascript,postMessage(1)');
postMessage(2);
/*
-->
<!doctype html>
<title>postMessage in importScripts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  var i = 0;
  worker.onmessage = this.step_func(function(e) {
    i++;
    assert_equals(e.data, i);
    if (i == 2)
      this.done();
  });
});
</script>
<!--
*/
//-->
